Split API handlers into domain modules
- Move application, operation, and system routes into dedicated modules
- Keep router wiring in `api/mod.rs` while preserving existing endpoints
